save: Use `dd` with the `notrunc` option
This will stop the overall truncation of the target file.
We need to do this because dd, like other coreutils, already truncates the file
on open(). In case we can't store the backup file afterwards we would end up in
a truncated file for which the user has no write permission by default.