Do not recalculate locations, just store them.
The code in `DoTextEvent` is quite hard to understand.
Why do we clamp an existing event here? Because we recalculate the end location
which should have been assigned before in `ExecuteTextEvent`.
The end locations for all `TextEventInsert` are incorrectly being stored in the undo stack and are hot-fixed while being processed.

This also fixes potential issues with `TextEventReplace` in regards to multi-line strings.