Add a new colorscheme group for selections within highlighted search results
micro allows to set the color for selections and for search results.
However, if your color scheme defines a color for highlights that is very similar to your general background color, and you perform a search, the currently selected text will hardly stand out.

So add a new color-group analogous to `statusline`, `statusline.inactive`, `statusline.suggestions`