plugin: expose NewTabFromBuffer to Lua
There is no really good way to create a new tab
that’s not bound to a specific file on disk without
this function. Now you can create a new buffer from string,
pass it to this function and add the resulting tab to the tabs list.

It’s still not particularly ergonomic, but at least it’s possible
with a few lines of boilerplate.